+++ server/lib/src/recordings.dart
@@ -1,6 +1,8 @@
 import 'package:aqueduct/aqueduct.dart';
 import 'package:musicus_database/musicus_database.dart';
 
+import 'auth.dart';
+
 class RecordingsController extends ResourceController {
   final Database db;
 
@@ -19,6 +21,16 @@ class RecordingsController extends ResourceController {
   @Operation.put('id')
   Future<Response> putRecording(
       @Bind.path('id') int id, @Bind.body() Map<String, dynamic> json) async {
+    if (await db.recordingById(id).getSingle() != null) {
+      if (!request.mayEdit) {
+        return Response.forbidden();
+      }
+    } else {
+      if (!request.mayUpload) {
+        return Response.forbidden();
+      }
+    }
+    
     final recordingInfo = RecordingInfo.fromJson(json);
     await db.updateRecording(recordingInfo);
     
@@ -27,6 +39,10 @@ class RecordingsController extends ResourceController {
 
   @Operation.delete('id')
   Future<Response> deleteRecording(@Bind.path('id') int id) async {
+    if (!request.mayDelete) {
+      return Response.forbidden();
+    }
+    
     await db.deleteRecording(id);
     return Response.ok(null);
   }
